Normal user can capture traffic.
While installing wireshark-common, dpkg recommends to allow non root
users to capture traffic. To do so, we answer with "yes" the question
'wireshark-common/install-setuid'.

"debconf" only sets the default option to that question, does not submit
it. Therefore, we must set it _before_ we install any package.

On the other hand, it is not needed to create the "wireshark" group
manually because dpkg will do it for us. Nevertheless, we need to add
our user to the group to enable this capability for it.